+++ src/stores/typedStoreHelpers.ts
@@ -14,6 +14,7 @@ const TEST_START_OPTIONS = makePrendyOptions({
   zoomLevels: { default: 1.1, max: 2 },
   walkSpeed: 10,
   animationSpeed: 1,
+  timeSpeed: 1,
   headHeightOffsets: { walker: 2.75 },
   doorsInfo: {
     street: {
@@ -264,9 +265,11 @@ export type PrendyOptionsUntyped = {
   };
   walkSpeed: number;
   animationSpeed: number; // 1.75 for rodont
+  gameTimeSpeed: number;
   headHeightOffsets: Record<string, number>; // maybe eventually move to being automatic by finding a bone with"neck" in its name
   doorsInfo?: Partial<Record<string, Partial<Record<string, ToNewOptionUntyped>>>>;
   modelNamesByPlace: Record<string, string[]>;
+  //
   // NOTE could add charactersWithSpeechBubbles (or dollsWithSpeechBubbles , or another way to define speechBubbles outside of characters)
   hasInteracting?: boolean;
   hasJumping?: boolean;
@@ -304,6 +307,7 @@ export type PrendyOptionsGeneric<
   };
   walkSpeed: number;
   animationSpeed: number;
+  timeSpeed: number;
   headHeightOffsets: Record<string, number>; // maybe eventually move to being automatic by finding a bone with"neck" in its name
   doorsInfo?: Partial<
     Record<PlaceName, Partial<Record<string, ToPlaceOption<AnyCameraName, AnySegmentName, PlaceName, AnySpotName>>>>